Add separete inviteLink element
